improve sanitation

+++ admin/class-accordion-slider-admin.php
@@ -161,7 +161,7 @@ class BQW_Accordion_Slider_Admin {
 			wp_enqueue_script( $this->plugin_slug . '-easing-script', plugins_url( 'public/assets/libs/easing/jquery.easing.1.3.min.js', dirname( __FILE__ ) ), array(), BQW_Accordion_Slider::VERSION );
 			wp_enqueue_script( $this->plugin_slug . '-video-js-script', plugins_url( 'public/assets/libs/video-js/video.js', dirname( __FILE__ ) ), array(), BQW_Accordion_Slider::VERSION );
 
-			$id = isset( $_GET['id'] ) ? $_GET['id'] : -1;
+			$id = isset( $_GET['id'] ) ? intval( $_GET['id'] ) : -1;
 
 			wp_localize_script( $this->plugin_slug . '-admin-script', 'as_js_vars', array(
 				'admin' => admin_url( 'admin.php' ),
@@ -255,7 +255,7 @@ class BQW_Accordion_Slider_Admin {
 	 */
 	public function render_accordion_page() {
 		if ( isset( $_GET['id'] ) && isset( $_GET['action'] ) && $_GET['action'] === 'edit' ) {
-			$accordion = $this->plugin->get_accordion( $_GET['id'] );
+			$accordion = $this->plugin->get_accordion( intval( $_GET['id'] ) );
 
 			if ( $accordion !== false ) {
 				$accordion_id = $accordion['id'];
@@ -316,7 +316,7 @@ class BQW_Accordion_Slider_Admin {
 		if ( isset( $_POST['plugin_settings_update'] ) ) {
 			check_admin_referer( 'plugin-settings-update', 'plugin-settings-nonce' );
 
-			if ( isset( $_POST['load_stylesheets'] ) ) {
+			if ( isset( $_POST['load_stylesheets'] ) && array_key_exists( $_POST['load_stylesheets'] , $plugin_settings['load_stylesheets']['available_values'] ) ) {
 				$load_stylesheets = $_POST['load_stylesheets'];
 				update_option( 'accordion_slider_load_stylesheets', $load_stylesheets );
 			}
@@ -330,7 +330,7 @@ class BQW_Accordion_Slider_Admin {
 			}
 
 			if ( isset( $_POST['cache_expiry_interval'] ) ) {
-				$cache_expiry_interval = $_POST['cache_expiry_interval'];
+				$cache_expiry_interval = intval( $_POST['cache_expiry_interval'] );
 				update_option( 'accordion_slider_cache_expiry_interval', $cache_expiry_interval );
 			}
 
@@ -358,7 +358,7 @@ class BQW_Accordion_Slider_Admin {
 				update_option( 'accordion_slider_hide_image_size_warning', false );
 			}
 
-			if ( isset( $_POST['access'] ) ) {
+			if ( isset( $_POST['access'] ) && array_key_exists( $_POST['access'], $plugin_settings['access']['available_values'] ) ) {
 				$access = $_POST['access'];
 				update_option( 'accordion_slider_access', $access );
 			}
@@ -385,13 +385,13 @@ class BQW_Accordion_Slider_Admin {
 	 */
 	public function ajax_get_accordion_data() {
 		$nonce = $_GET['nonce'];
-		$id = $_GET['id'];
+		$id = intval( $_GET['id'] );
 
 		if ( ! wp_verify_nonce( $nonce, 'load-accordion-data' . $id ) ) {
 			die( 'This action was stopped for security purposes.' );
 		}
 
-		$accordion = $this->get_accordion_data( $_GET['id'] );
+		$accordion = $this->get_accordion_data( $id );
 
 		echo json_encode( $accordion );
 
@@ -420,10 +420,12 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.0.0
 	 */
 	public function ajax_save_accordion() {
-		$accordion_data = json_decode( stripslashes( $_POST['data'] ), true );
-		$nonce = $accordion_data['nonce'];
-		$id = intval( $accordion_data['id'] );
-		$action = $accordion_data['action'];
+		$data = json_decode( stripslashes( $_POST['data'] ), true );
+		$nonce = $data['nonce'];
+		$action = $data['action'];
+ 
+		$accordion_data = BQW_Accordion_Slider_Validation::validate_accordion_slider_data( $data );
+		$id = $accordion_data['id'];
 
 		if ( ! wp_verify_nonce( $nonce, 'save-accordion' . $id ) ) {
 			die( 'This action was stopped for security purposes.' );
@@ -432,7 +434,7 @@ class BQW_Accordion_Slider_Admin {
 		$accordion_id = $this->save_accordion( $accordion_data );
 
 		if ( $action === 'save' ) {
-			echo $accordion_id;
+			echo json_encode( $accordion_id );
 		} else if ( $action === 'import' ) {
 			$accordion_name = $accordion_data['name'];
 			$accordion_created = date( 'm-d-Y' );
@@ -462,7 +464,7 @@ class BQW_Accordion_Slider_Admin {
 	public function save_accordion( $accordion_data ) {
 		global $wpdb;
 
-		$id = intval( $accordion_data['id'] );
+		$id = $accordion_data['id'];
 		$panels_data = $accordion_data['panels'];
 
 		if ( $id === -1 ) {
@@ -551,8 +553,7 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.0.0
 	 */
 	public function ajax_preview_accordion() {
-		$accordion = json_decode( stripslashes( $_POST['data'] ), true );
-		$accordion_name = $accordion['name'];
+		$accordion = BQW_Accordion_Slider_Validation::validate_accordion_slider_data( json_decode( stripslashes( $_POST['data'] ), true ) );
 		$accordion_output = $this->plugin->output_accordion( $accordion, false ) . $this->plugin->get_inline_scripts();
 
 		include( 'views/preview-window.php' );
@@ -573,7 +574,7 @@ class BQW_Accordion_Slider_Admin {
 	 */
 	public function ajax_duplicate_accordion() {
 		$nonce = $_POST['nonce'];
-		$original_accordion_id = $_POST['id'];
+		$original_accordion_id = intval( $_POST['id'] );
 
 		if ( ! wp_verify_nonce( $nonce, 'duplicate-accordion' . $original_accordion_id ) ) {
 			die( 'This action was stopped for security purposes.' );
@@ -707,7 +708,7 @@ class BQW_Accordion_Slider_Admin {
 	 * AJAX call for adding multiple or a single panel.
 	 *
 	 * If it receives any data, it tries to create multiple
-	 * panels by padding the data that was received, and if
+	 * panels by passing the data that was received, and if
 	 * it doesn't receive any data it tries to create a
 	 * single panel.
 	 *
@@ -715,7 +716,7 @@ class BQW_Accordion_Slider_Admin {
 	 */
 	public function ajax_add_panels() {
 		if ( isset( $_POST['data'] ) ) {
-			$panels_data = json_decode( stripslashes( $_POST['data'] ), true );
+			$panels_data = BQW_Accordion_Slider_Validation::validate_accordion_slider_panels( json_decode( stripslashes( $_POST['data'] ), true ) );
 
 			foreach ( $panels_data as $panel_data ) {
 				$this->create_panel( $panel_data );
@@ -739,8 +740,8 @@ class BQW_Accordion_Slider_Admin {
 	public function ajax_load_background_image_editor() {
 		$panel_default_settings = BQW_Accordion_Slider_Settings::getPanelSettings();
 
-		$data = json_decode( stripslashes( $_POST['data'] ), true );
-		$content_type = isset( $_POST['content_type'] ) ? $_POST['content_type'] : $panel_default_settings['content_type']['default_value'];
+		$data = reset( BQW_Accordion_Slider_Validation::validate_accordion_slider_panels( array( json_decode( stripslashes( $_POST['data'] ), true ) ) ) );
+		$content_type = isset( $_POST['content_type'] ) && array_key_exists( $_POST['content_type'], $panel_default_settings['content_type']['available_values'] ) ? $_POST['content_type'] : $panel_default_settings['content_type']['default_value'];
 		$content_class = $content_type === 'custom' ? 'custom' : 'dynamic';
 
 		include( 'views/background-image-editor.php' );
@@ -755,9 +756,29 @@ class BQW_Accordion_Slider_Admin {
 	 */
 	public function ajax_load_html_editor() {
 		$panel_default_settings = BQW_Accordion_Slider_Settings::getPanelSettings();
+		global $allowedposttags;
 
-		$html_content = $_POST['data'];
-		$content_type = isset( $_POST['content_type'] ) ? $_POST['content_type'] : $panel_default_settings['content_type']['default_value'];
+		$allowed_html = array_merge(
+			$allowedposttags,
+			array(
+				'iframe' => array(
+					'src' => true,
+					'width' => true,
+					'height' => true,
+					'allow' => true,
+					'allowfullscreen' => true,
+					'class' => true,
+					'id' => true
+				),
+				'source' => array(
+					'src' => true,
+					'type' => true
+				)
+			)
+		);
+
+		$html_content = wp_kses( $_POST['data'], $allowed_html );
+		$content_type = isset( $_POST['content_type'] ) && array_key_exists( $_POST['content_type'], $panel_default_settings['content_type']['available_values'] ) ? $_POST['content_type'] : $panel_default_settings['content_type']['default_value'];
 
 		include( 'views/html-editor.php' );
 
@@ -773,9 +794,9 @@ class BQW_Accordion_Slider_Admin {
 		$panel_default_settings = BQW_Accordion_Slider_Settings::getPanelSettings();
 		$layer_default_settings = BQW_Accordion_Slider_Settings::getLayerSettings();
 
-		$layers = json_decode( stripslashes( $_POST['data'] ), true );
-		$content_type = isset( $_POST['content_type'] ) ? $_POST['content_type'] : $panel_default_settings['content_type']['default_value'];
-		
+		$layers = BQW_Accordion_Slider_Validation::validate_panel_layers( json_decode( stripslashes( $_POST['data'] ), true ) );
+		$content_type = isset( $_POST['content_type'] ) && array_key_exists( $_POST['content_type'], $panel_default_settings['content_type']['available_values'] ) ? $_POST['content_type'] : $panel_default_settings['content_type']['default_value'];
+	
 		include( 'views/layers-editor.php' );
 
 		die();
@@ -790,37 +811,58 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.0.0
 	 */
 	public function ajax_add_layer_settings() {
+		$layer_default_settings = BQW_Accordion_Slider_Settings::getLayerSettings();
 		$layer = array();
-		$layer_id = $_POST['id'];
-		$layer_type = $_POST['type'];
+		$layer_id = intval( $_POST['id'] );
+		$layer_type = isset( $_POST['type'] ) && array_key_exists( $_POST['type'], $layer_default_settings['type']['available_values'] ) ? $_POST['type'] : $layer_default_settings['type']['default_value'];
 		$layer_settings;
+		global $allowedposttags;
 
 		if ( isset( $_POST['settings'] ) ) {
-			$layer_settings = json_decode( stripslashes( $_POST['settings'] ), true );
+			$layer_settings = BQW_Accordion_Slider_Validation::validate_layer_settings( json_decode( stripslashes( $_POST['settings'] ), true ) );
 		}
 
 		if ( isset( $_POST['text'] ) ) {
-			$layer['text'] = $_POST['text'];
+			$allowed_html = array_merge(
+				$allowedposttags,
+				array(
+					'iframe' => array(
+						'src' => true,
+						'width' => true,
+						'height' => true,
+						'allow' => true,
+						'allowfullscreen' => true,
+						'class' => true,
+						'id' => true
+					),
+					'source' => array(
+						'src' => true,
+						'type' => true
+					)
+				)
+			);
+
+			$layer['text'] = wp_kses( $_POST['text'], $allowed_html );
 		}
 
 		if ( isset( $_POST['heading_type'] ) ) {
-			$layer['heading_type'] = $_POST['heading_type'];
+			$layer['heading_type'] = array_key_exists( $_POST['heading_type'], $layer_default_settings['heading_type']['available_values'] ) ? $_POST['heading_type'] : $layer_default_settings['heading_type']['default_value'];
 		}
 
 		if ( isset( $_POST['image_source'] ) ) {
-			$layer['image_source'] = $_POST['image_source'];
+			$layer['image_source'] = sanitize_text_field( $_POST['image_source'] );
 		}
 
 		if ( isset( $_POST['image_alt'] ) ) {
-			$layer['image_alt'] = $_POST['image_alt'];
+			$layer['image_alt'] = sanitize_text_field( $_POST['image_alt'] );
 		}
 
 		if ( isset( $_POST['image_link'] ) ) {
-			$layer['image_link'] = $_POST['image_link'];
+			$layer['image_link'] = sanitize_text_field( $_POST['image_link'] );
 		}
 
 		if ( isset( $_POST['image_retina'] ) ) {
-			$layer['image_retina'] = $_POST['image_retina'];
+			$layer['image_retina'] = sanitize_text_field( $_POST['image_retina'] );
 		}
 
 		$layer_default_settings = BQW_Accordion_Slider_Settings::getLayerSettings();
@@ -836,11 +878,9 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.0.0
 	 */
 	public function ajax_load_settings_editor() {
-		$panel_settings = json_decode( stripslashes( $_POST['data'] ), true );
-
 		$panel_default_settings = BQW_Accordion_Slider_Settings::getPanelSettings();
-
-		$content_type = isset( $panel_settings['content_type'] ) ? $panel_settings['content_type'] : $panel_default_settings['content_type']['default_value'];
+		$panel_settings = BQW_Accordion_Slider_Validation::validate_panel_settings( json_decode( stripslashes( $_POST['data'] ), true ) );
+		$content_type = isset( $panel_settings['content_type'] ) && array_key_exists( $panel_settings['content_type'], $panel_default_settings['content_type']['available_values'] ) ? $panel_settings['content_type'] : $panel_default_settings['content_type']['default_value'];
 
 		include( 'views/settings-editor.php' );
 
@@ -878,7 +918,9 @@ class BQW_Accordion_Slider_Admin {
 	 */
 	public function load_content_type_settings( $type, $panel_settings = NULL ) {
 		$panel_default_settings = BQW_Accordion_Slider_Settings::getPanelSettings();
-
+		$type = isset( $_POST['type'] ) && array_key_exists( $_POST['type'], $panel_default_settings['content_type']['available_values'] ) ? $_POST['type'] : $panel_default_settings['content_type']['default_value'];
+		$panel_settings = BQW_Accordion_Slider_Validation::validate_panel_settings( json_decode( stripslashes( $_POST['data'] ), true ) );
+ 
 		if ( $type === 'posts' ) {
 			$post_names = $this->get_post_names();
 
@@ -941,7 +983,12 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.0.0
 	 */
 	public function ajax_get_taxonomies() {
-		$post_names = json_decode( stripslashes( $_GET['post_names'] ), true );
+		$post_names_raw = json_decode( stripslashes( $_GET['post_names'] ), true );
+		$post_names = array();
+ 
+		foreach ( $post_names_raw as $post_name ) {
+			array_push( $post_names, sanitize_text_field( $post_name ) );
+		}
 
 		echo json_encode( $this->get_taxonomies_for_posts( $post_names ) );
 
@@ -1026,7 +1073,7 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.0.0
 	 */
 	public function ajax_add_breakpoint() {
-		$width = $_GET['data'];
+		$width = floatval( $_GET['data'] );
 
 		include( 'views/breakpoint.php' );
 
@@ -1039,7 +1086,7 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.0.0
 	 */
 	public function ajax_add_breakpoint_setting() {
-		$setting_name = $_GET['data'];
+		$setting_name = sanitize_text_field( $_GET['data'] );
 
 		echo $this->create_breakpoint_setting( $setting_name, false );
